Unit test updates to use proper interpolation value

+++ contracts/test/unit/peg/UnitTestPegStabilityModule.t.sol
@@ -62,7 +62,10 @@ contract UnitTestPegStabilityModule is Test {
         underlyingToken = IERC20(address(new MockERC20()));
         core = getCoreV2();
         volt = core.volt();
-        (oracle, ) = getLocalOracleSystem(address(core), voltFloorPrice);
+        (oracle, ) = getLocalOracleSystem(
+            address(core),
+            uint112(voltFloorPrice)
+        );
         lock = IGlobalReentrancyLock(
             address(new GlobalReentrancyLock(address(core)))
         );
